requirements.txt
ttkbootstrap
Pillow